+++ capa/features/extractors/dnfile/insn.py
@@ -17,20 +17,25 @@ from dncil.cil.opcode import OpCodes
 from dncil.cil.instruction import Instruction
 
 import capa.features.extractors.helpers
-from capa.features.insn import API, Number
-from capa.features.common import Class, String, Feature, Namespace, Characteristic
+from capa.features.insn import API, Number, Property
+from capa.features.common import Class, String, Feature, Namespace, FeatureAccess, Characteristic
 from capa.features.address import Address
 from capa.features.extractors.base_extractor import BBHandle, InsnHandle, FunctionHandle
 from capa.features.extractors.dnfile.helpers import (
-    DnClass,
-    DnMethod,
+    DnType,
     DnUnmanagedMethod,
+    get_dotnet_fields,
     resolve_dotnet_token,
+    get_dotnet_properties,
     read_dotnet_user_string,
     get_dotnet_managed_imports,
     get_dotnet_managed_methods,
     get_dotnet_unmanaged_imports,
 )
+
+METHODDEF_TABLE = dnfile.mdtable.MethodDef.number
+MEMBERREF_TABLE = dnfile.mdtable.MemberRef.number
+FIELD_TABLE = dnfile.mdtable.Field.number
 
 
 def get_managed_imports(ctx: Dict) -> Dict:
@@ -57,16 +62,32 @@ def get_methods(ctx: Dict) -> Dict:
     return ctx["methods_cache"]
 
 
-def get_callee(ctx: Dict, token: int) -> Union[DnMethod, DnUnmanagedMethod, None]:
+def get_callee(ctx: Dict, token: int) -> Union[DnType, DnUnmanagedMethod, None]:
     """map dotnet token to un/managed method"""
-    callee: Union[DnMethod, DnUnmanagedMethod, None] = get_managed_imports(ctx).get(token, None)
-    if not callee:
+    callee: Union[DnType, DnUnmanagedMethod, None] = get_managed_imports(ctx).get(token, None)
+    if callee is None:
         # we must check unmanaged imports before managed methods because we map forwarded managed methods
         # to their unmanaged imports; we prefer a forwarded managed method be mapped to its unmanaged import for analysis
         callee = get_unmanaged_imports(ctx).get(token, None)
-        if not callee:
+        if callee is None:
             callee = get_methods(ctx).get(token, None)
     return callee
+
+
+def get_properties(ctx: Dict) -> Dict:
+    if "properties_cache" not in ctx:
+        ctx["properties_cache"] = {}
+        for prop in get_dotnet_properties(ctx["pe"]):
+            ctx["properties_cache"][prop.token] = prop
+    return ctx["properties_cache"]
+
+
+def get_fields(ctx: Dict) -> Dict:
+    if "fields_cache" not in ctx:
+        ctx["fields_cache"] = {}
+        for field in get_dotnet_fields(ctx["pe"]):
+            ctx["fields_cache"][field.token] = field
+    return ctx["fields_cache"]
 
 
 def extract_insn_api_features(fh: FunctionHandle, bh, ih: InsnHandle) -> Iterator[Tuple[Feature, Address]]:
@@ -76,49 +97,148 @@ def extract_insn_api_features(fh: FunctionHandle, bh, ih: InsnHandle) -> Iterato
     if insn.opcode not in (OpCodes.Call, OpCodes.Callvirt, OpCodes.Jmp, OpCodes.Calli):
         return
 
-    callee: Union[DnMethod, DnUnmanagedMethod, None] = get_callee(fh.ctx, insn.operand.value)
+    callee: Union[DnType, DnUnmanagedMethod, None] = get_callee(fh.ctx, insn.operand.value)
     if callee is None:
         return
 
-    if isinstance(callee, DnUnmanagedMethod):
-        # like kernel32.CreateFileA
-        for name in capa.features.extractors.helpers.generate_symbols(callee.modulename, callee.methodname):
-            yield API(name), ih.address
-    else:
+    if isinstance(callee, DnType):
+        if callee.member.startswith(("get_", "set_")):
+            if insn.operand.table == METHODDEF_TABLE:
+                # check if the method belongs to the MethodDef table and whether it is used to access a property
+                if get_properties(fh.ctx).get(insn.operand.value, None) is not None:
+                    return
+            elif insn.operand.table == MEMBERREF_TABLE:
+                # if the method belongs to the MemberRef table, we assume it is used to access a property
+                return
+
         # like System.IO.File::Delete
         yield API(str(callee)), ih.address
+
+    else:
+        # like kernel32.CreateFileA
+        for name in capa.features.extractors.helpers.generate_symbols(callee.module, callee.method):
+            yield API(name), ih.address
+
+
+def extract_insn_property_features(fh: FunctionHandle, bh, ih: InsnHandle) -> Iterator[Tuple[Feature, Address]]:
+    """parse instruction property features"""
+    insn: Instruction = ih.inner
+
+    name: Optional[str] = None
+    access: Optional[str] = None
+
+    if insn.opcode in (OpCodes.Call, OpCodes.Callvirt, OpCodes.Jmp, OpCodes.Calli):
+        if insn.operand.table == METHODDEF_TABLE:
+            # check if the method belongs to the MethodDef table and whether it is used to access a property
+            prop = get_properties(fh.ctx).get(insn.operand.value, None)
+            if prop is not None:
+                name = str(prop)
+                access = prop.access
+
+        elif insn.operand.table == MEMBERREF_TABLE:
+            # if the method belongs to the MemberRef table, we assume it is used to access a property
+            row: Any = resolve_dotnet_token(fh.ctx["pe"], insn.operand)
+            if row is None:
+                return
+            if not isinstance(row.Class.row, (dnfile.mdtable.TypeRefRow, dnfile.mdtable.TypeDefRow)):
+                return
+            if not row.Name.startswith(("get_", "set_")):
+                return
+
+            name = DnType.format_name(
+                row.Class.row.TypeName, namespace=row.Class.row.TypeNamespace, member=row.Name[4:]
+            )
+            if row.Name.startswith("get_"):
+                access = FeatureAccess.READ
+            elif row.Name.startswith("set_"):
+                access = FeatureAccess.WRITE
+
+    elif insn.opcode in (OpCodes.Ldfld, OpCodes.Ldflda, OpCodes.Ldsfld, OpCodes.Ldsflda):
+        if insn.operand.table == FIELD_TABLE:
+            # determine whether the operand is a field by checking if it belongs to the Field table
+            read_field: Optional[DnType] = get_fields(fh.ctx).get(insn.operand.value, None)
+            if read_field:
+                name = str(read_field)
+                access = FeatureAccess.READ
+
+    elif insn.opcode in (OpCodes.Stfld, OpCodes.Stsfld):
+        if insn.operand.table == FIELD_TABLE:
+            # determine whether the operand is a field by checking if it belongs to the Field table
+            write_field: Optional[DnType] = get_fields(fh.ctx).get(insn.operand.value, None)
+            if write_field:
+                name = str(write_field)
+                access = FeatureAccess.WRITE
+
+    if name is not None:
+        if access is not None:
+            yield Property(name, access=access), ih.address
+        yield Property(name), ih.address
 
 
 def extract_insn_class_features(fh: FunctionHandle, bh, ih: InsnHandle) -> Iterator[Tuple[Class, Address]]:
     """parse instruction class features"""
-    if ih.inner.opcode not in (OpCodes.Call, OpCodes.Callvirt, OpCodes.Jmp, OpCodes.Calli):
+    if ih.inner.opcode not in (
+        OpCodes.Call,
+        OpCodes.Callvirt,
+        OpCodes.Jmp,
+        OpCodes.Calli,
+        OpCodes.Ldfld,
+        OpCodes.Ldflda,
+        OpCodes.Ldsfld,
+        OpCodes.Ldsflda,
+        OpCodes.Stfld,
+        OpCodes.Stsfld,
+    ):
         return
 
-    row: Any = resolve_dotnet_token(fh.ctx["pe"], Token(ih.inner.operand.value))
+    row: Any = resolve_dotnet_token(fh.ctx["pe"], ih.inner.operand)
+    if isinstance(row, dnfile.mdtable.MemberRefRow):
+        if isinstance(row.Class.row, (dnfile.mdtable.TypeRefRow, dnfile.mdtable.TypeDefRow)):
+            yield Class(DnType.format_name(row.Class.row.TypeName, namespace=row.Class.row.TypeNamespace)), ih.address
 
-    if not isinstance(row, dnfile.mdtable.MemberRefRow):
-        return
-    if not isinstance(row.Class.row, (dnfile.mdtable.TypeRefRow, dnfile.mdtable.TypeDefRow)):
-        return
+    elif isinstance(row, dnfile.mdtable.MethodDefRow):
+        callee: Union[DnType, DnUnmanagedMethod, None] = get_callee(fh.ctx, ih.inner.operand.value)
+        if isinstance(callee, DnType):
+            yield Class(DnType.format_name(callee.class_, namespace=callee.namespace)), ih.address
 
-    yield Class(DnClass.format_name(row.Class.row.TypeNamespace, row.Class.row.TypeName)), ih.address
+    elif isinstance(row, dnfile.mdtable.FieldRow):
+        field: Optional[DnType] = get_fields(fh.ctx).get(ih.inner.operand.value, None)
+        if field is not None:
+            yield Class(DnType.format_name(field.class_, namespace=field.namespace)), ih.address
 
 
 def extract_insn_namespace_features(fh: FunctionHandle, bh, ih: InsnHandle) -> Iterator[Tuple[Namespace, Address]]:
     """parse instruction namespace features"""
-    if ih.inner.opcode not in (OpCodes.Call, OpCodes.Callvirt, OpCodes.Jmp, OpCodes.Calli):
+    if ih.inner.opcode not in (
+        OpCodes.Call,
+        OpCodes.Callvirt,
+        OpCodes.Jmp,
+        OpCodes.Calli,
+        OpCodes.Ldfld,
+        OpCodes.Ldflda,
+        OpCodes.Ldsfld,
+        OpCodes.Ldsflda,
+        OpCodes.Stfld,
+        OpCodes.Stsfld,
+    ):
         return
 
     row: Any = resolve_dotnet_token(fh.ctx["pe"], Token(ih.inner.operand.value))
 
-    if not isinstance(row, dnfile.mdtable.MemberRefRow):
-        return
-    if not isinstance(row.Class.row, (dnfile.mdtable.TypeRefRow, dnfile.mdtable.TypeDefRow)):
-        return
-    if not row.Class.row.TypeNamespace:
-        return
+    if isinstance(row, dnfile.mdtable.MemberRefRow):
+        if isinstance(row.Class.row, (dnfile.mdtable.TypeRefRow, dnfile.mdtable.TypeDefRow)):
+            if row.Class.row.TypeNamespace:
+                yield Namespace(row.Class.row.TypeNamespace), ih.address
 
-    yield Namespace(row.Class.row.TypeNamespace), ih.address
+    elif isinstance(row, dnfile.mdtable.MethodDefRow):
+        callee: Union[DnType, DnUnmanagedMethod, None] = get_callee(fh.ctx, ih.inner.operand.value)
+        if isinstance(callee, DnType) and callee.namespace is not None:
+            yield Namespace(callee.namespace), ih.address
+
+    elif isinstance(row, dnfile.mdtable.FieldRow):
+        field: Optional[DnType] = get_fields(fh.ctx).get(ih.inner.operand.value, None)
+        if field is not None:
+            yield Namespace(field.namespace), ih.address
 
 
 def extract_insn_number_features(fh, bh, ih: InsnHandle) -> Iterator[Tuple[Feature, Address]]:
@@ -174,6 +294,7 @@ def extract_features(fh: FunctionHandle, bbh: BBHandle, ih: InsnHandle) -> Itera
 
 INSTRUCTION_HANDLERS = (
     extract_insn_api_features,
+    extract_insn_property_features,
     extract_insn_number_features,
     extract_insn_string_features,
     extract_insn_namespace_features,