+++ tests/fixtures.py
@@ -36,6 +36,7 @@ from capa.features.common import (
     Arch,
     Format,
     Feature,
+    FeatureAccess,
 )
 from capa.features.address import Address
 from capa.features.extractors.base_extractor import BBHandle, InsnHandle, FunctionHandle
@@ -279,6 +280,10 @@ def get_data_path_by_name(name):
         return os.path.join(CD, "data", "dotnet", "1c444ebeba24dcba8628b7dfe5fec7c6.exe_")
     elif name.startswith("_692f"):
         return os.path.join(CD, "data", "dotnet", "692f7fd6d198e804d6af98eb9e390d61.exe_")
+    elif name.startswith("_0953c"):
+        return os.path.join(CD, "data", "0953cc3b77ed2974b09e3a00708f88de931d681e2d0cb64afbaf714610beabe6.exe_")
+    elif name.startswith("_039a6"):
+        return os.path.join(CD, "data", "039a6336d0802a2255669e6867a5679c7eb83313dbc61fb1c7232147379bd304.exe_")
     else:
         raise ValueError("unexpected sample fixture: %s" % name)
 
@@ -758,6 +763,106 @@ FEATURE_PRESENCE_TESTS_DOTNET = sorted(
             True,
         ),
         ("_1c444", "function=0x1F68, bb=0x1F68, insn=0x1FF9", capa.features.insn.API("FromHbitmap"), False),
+        (
+            "_1c444",
+            "token=0x600002B",
+            capa.features.insn.Property("System.IO.FileInfo::Length", access=FeatureAccess.READ),
+            True,
+        ),  # MemberRef method
+        (
+            "_1c444",
+            "token=0x600002B",
+            capa.features.insn.Property("System.IO.FileInfo::Length"),
+            True,
+        ),  # MemberRef method
+        (
+            "_1c444",
+            "token=0x6000081",
+            capa.features.insn.API("System.Diagnostics.Process::Start"),
+            True,
+        ),  # MemberRef method
+        (
+            "_1c444",
+            "token=0x6000081",
+            capa.features.insn.Property(
+                "System.Diagnostics.ProcessStartInfo::UseShellExecute", access=FeatureAccess.WRITE
+            ),  # MemberRef method
+            True,
+        ),
+        (
+            "_1c444",
+            "token=0x6000081",
+            capa.features.insn.Property(
+                "System.Diagnostics.ProcessStartInfo::WorkingDirectory", access=FeatureAccess.WRITE
+            ),  # MemberRef method
+            True,
+        ),
+        (
+            "_1c444",
+            "token=0x6000081",
+            capa.features.insn.Property(
+                "System.Diagnostics.ProcessStartInfo::FileName", access=FeatureAccess.WRITE
+            ),  # MemberRef method
+            True,
+        ),
+        (
+            "_1c444",
+            "token=0x6000087",
+            capa.features.insn.Property("Sockets.MySocket::reConnectionDelay", access=FeatureAccess.WRITE),  # Field
+            True,
+        ),
+        (
+            "_1c444",
+            "token=0x600008A",
+            capa.features.insn.Property("Sockets.MySocket::isConnected", access=FeatureAccess.WRITE),  # Field
+            True,
+        ),
+        (
+            "_1c444",
+            "token=0x600008A",
+            capa.features.insn.Property("Sockets.MySocket::onConnected", access=FeatureAccess.READ),  # Field
+            True,
+        ),
+        (
+            "_0953c",
+            "token=0x6000004",
+            capa.features.insn.Property("System.Diagnostics.Debugger::IsAttached", access=FeatureAccess.READ),
+            True,
+        ),  # MemberRef method
+        (
+            "_692f",
+            "token=0x6000006",
+            capa.features.insn.Property(
+                "System.Management.Automation.PowerShell::Streams", access=FeatureAccess.READ
+            ),  # MemberRef method
+            False,
+        ),
+        (
+            "_039a6",
+            "token=0x6000007",
+            capa.features.insn.API("System.Reflection.Assembly::Load"),
+            True,
+        ),
+        (
+            "_039a6",
+            "token=0x600001D",
+            capa.features.insn.Property("StagelessHollow.Arac::Marka", access=FeatureAccess.READ),  # MethodDef method
+            True,
+        ),
+        (
+            "_039a6",
+            "token=0x600001C",
+            capa.features.insn.Property("StagelessHollow.Arac::Marka", access=FeatureAccess.READ),  # MethodDef method
+            False,
+        ),
+        (
+            "_039a6",
+            "token=0x6000023",
+            capa.features.insn.Property(
+                "System.Runtime.CompilerServices.AsyncTaskMethodBuilder::Task", access=FeatureAccess.READ
+            ),  # MemberRef method
+            False,
+        ),
     ],
     # order tests by (file, item)
     # so that our LRU cache is most effective.
@@ -904,3 +1009,13 @@ def _1c444_dotnetfile_extractor():
 @pytest.fixture
 def _692f_dotnetfile_extractor():
     return get_dnfile_extractor(get_data_path_by_name("_692f"))
+
+
+@pytest.fixture
+def _0953c_dotnetfile_extractor():
+    return get_dnfile_extractor(get_data_path_by_name("_0953c"))
+
+
+@pytest.fixture
+def _039a6_dotnetfile_extractor():
+    return get_dnfile_extractor(get_data_path_by_name("_039a6"))