+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New Features
 - verify rule metadata format on load #1160 @mr-tz
+- extract property features from .NET PE files #1168 @anushkavirgaonkar
 
 ### Breaking Changes